Added support for command line parameter options:
1 -inputfile to start the app with a textfile which will be parsed.
a if this parameter has no value, the app switches to the default console input.

2 -exporter to use a custom exporter.
a if this parameter has no value, the user will get a list to choose from in the console.
b if the name of the exporter is not valid, the app switches to the default exporter, YUML.